requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astor==0.8.1
attrs==24.3.0
Automat==24.8.1
bert-score==0.3.13
bitsandbytes==0.45.0
blake3==1.0.2
Brotli
cachetools==5.5.0
certifi
charset-normalizer
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
colorful==0.5.6
comet_ml==3.49.9
compressed-tensors==0.8.1
configobj==5.0.9
constantly==23.10.4
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
einops==0.8.0
entmax==1.3
everett==3.1.0
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
googletrans==4.0.2
greenlet==3.2.1
grpcio==1.69.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
hyperframe==6.1.0
hyperlink==21.0.0
idna
importlib_metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
interegular==0.3.3
ipadic==1.0.0
jieba==0.42.1
Jinja2
jiter==0.8.2
joblib==1.4.2
jsonargparse==3.13.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.42
lark==1.2.2
lightning-utilities==0.11.9
linkify-it-py==2.0.3
lm-format-enforcer==0.10.9
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe
marshmallow==3.26.1
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mecab-python3==1.0.10
memray==1.15.0
mistral_common==1.5.1
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx
nltk==3.9.1
numpy==1.26.4
nvidia-ml-py==12.560.30
openai==1.59.7
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.1
py-cpuinfo==9.0.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks
pytest==8.3.4
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
sacrebleu==2.5.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.27.0
setuptools==75.8.0
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchmetrics==0.10.3
torchvision==0.20.1
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers.git@4b8d1f7fca7b3f043ddaf375b7c1417064b120c5
triton==3.1.0
Twisted==24.11.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions
tzdata==2024.2
uc-micro-py==1.0.3
unbabel-comet==2.2.4
urllib3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.1
vllm==0.6.6.post1
watchfiles==1.0.4
websockets==14.1
wheel==0.44.0
wrapt==1.17.1
wurlitzer==3.1.1
xformers==0.0.28.post3
xgrammar==0.1.9
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
